Refactor code to match coding guidelines and better documentation.
