+++ keywords.txt
@@ -567,6 +567,14 @@ P9813Spi2MhzMethod	KEYWORD1
 P9813Spi1MhzMethod	KEYWORD1
 P9813Spi500KhzMethod	KEYWORD1
 P9813SpiHzMethod	KEYWORD1
+Tlc5947Method	KEYWORD1
+Tlc5947Method16Bit	KEYWORD1
+Tlc5947Spi30MhzMethod	KEYWORD1
+Tlc5947Spi30MhzMethod16Bit	KEYWORD1
+Tlc5947Spi15MhzMethod	KEYWORD1
+Tlc5947Spi15MhzMethod16Bit	KEYWORD1
+Tlc5947SpiMethod	KEYWORD1
+Tlc5947SpiMethod16Bit	KEYWORD1
 NeoPixelAnimator	KEYWORD1
 AnimUpdateCallback	KEYWORD1
 AnimationParam	KEYWORD1
